Make process pixels part of a utilities
